pyproject.toml
[project]
name = "Luk-chan"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Gabriel Nobrega", email = "gabrieltkdnobrega63@gmail.com" },
]
dependencies = ["discord-py>=2.6.3", "pydantic-settings>=2.11.0", "dateparser>=1.2.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = { cmd = "python main.py", env_file.override = ".env" }

[dependency-groups]
dev = ["ruff>=0.13.2"]
